Package refactor - RelocationHelper*
