--- slimjar/src/main/java/io/github/slimjar/relocation/RelocationHelper.java
@@ -1,10 +1,0 @@
-package io.github.slimjar.relocation;
-
-import io.github.slimjar.resolver.data.Dependency;
-
-import java.io.File;
-import java.io.IOException;
-
-public interface RelocationHelper {
-    File relocate(final Dependency dependency, final File file) throws IOException;
-}
